add to db and generate text to file

+++ src/main/resources/sql/init.sql
@@ -1,13 +1,13 @@
 -- CREATE DATABASE smart_translator;
 
 CREATE TABLE IF NOT EXISTS words (
-    id BIGINT PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     text VARCHAR(50) NOT NULL,
     language VARCHAR(50) NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS translations (
-    id BIGINT PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     original_word_id INT NOT NULL REFERENCES words(id),
     translated_word_id INT NOT NULL REFERENCES words(id)
 );
